Consider Super_L and Super_R as modifiers
Without this patch, I can't use the meta key to assign shortcuts, as
Super_L and Super_R are not considered as modifiers, so when I press
meta (Super_L on my system), the shortcut is immediately accepted,
before I get the chance to press another key.

This patch requires the fix in
https://bugreports.qt-project.org/browse/QTBUG-38428
to be applied. With both patches, the QtQuick widgets to assign
shortcuts works for me also with the Meta key.

Suggested by David Edmundson in review request at
https://git.reviewboard.kde.org/r/118581/

If the upstream Qt fix changes, we'll have to readjust here, but at
least now we have a good chance of it working.

CCBUG:335316
REVIEW:118630