Add a plotter component
a Plotter can draw a graph of values arriving from an arbitrary number of data sources
to show their evoluton in time.
an example can be a plot of the network transfer speed or CPU temperature over time.
Multiple plots can be fitted in the same graph, either stacked or intersected.

REVIEW:121581