Add overload method for Formats::formatDuration taking int
It is currently not possible to use Q_INVOKABLE that takes
Q_ENUM/Q_FLAGS as a parameter that is defined in a different class. This
is mentioned in the docs and there's also a bug report here -
https://bugreports.qt.io/browse/QTBUG-20639

If the default parameter is removed from the formatDuration, Plasma will
complain with "Error: Unknown method parameter type
KFormat::DurationFormatOptions".

Until that one is fixed, we can either take an integer and cast it to
the flags or redeclare the enum from KFormats locally, but that's just
not good.

REVIEW: 122200