Add paintedWidth and paintedHeight to QPixmapItem
Like the corresponding properties in QML's Image, this patch introduces
paintedWidth and paintedHeight to QPixmapItem, to get information on the
size of the pixmap that's actually shown (depending on fillMode), as
compared to its actual size.

This brings QPixmapItem closer to the mode of usage of Image.

CHANGELOG: New properties paintedWidth and paintedHeight added to
QPixmapItem

REVIEW: 122413