Fix performance issue when dealing with DeclarativeMimeData
It was copying all the information upon construction, which didn't make
sense in the case of the DropArea because from DropArea we don't need to
modify the mime data.
This caused that whenever there was some elaborate QMimeData instance being
passed the whole process would freeze while copying. This was especially
evident when dealing with images.

Reviewed by Marco Martin