QML-only KCM base
This is a QObject-based fork of KCmodule, with the api slightly adaped
to be more declarative friendly.
it loads a QML file that then will be able to access its instance as
"kcm" or as the "KCM" attached property, like what happens in plasmoids
(for how attached proeprties work, unfortunately only properties of the
base class will be accessible).

REVIEW:122886