Same behavior of MouseArea to map coords of filtered child events
filtered mouse events and mouse events will have coordinates
mapped differently, having a touch and a mouse click on the same
place arriving with two different coordinates in the event filter.
for more details see
https://bugreports.qt.io/browse/QTBUG-46761
https://codereview.qt-project.org/#/c/114424/
It's unlikely this behavior to change in Qt as a change may
have strange sideeffects in current user code.

What we would want is coordinates arriving all mapped as they were
local coordinates of the MouseEventListener.
This makes it do like MouseArea, that maps from event->windowPos(),
making it actually bypassing the wrong pos() of the filtered event.

reviewed-by: mgraesslin@kde.org
The behavior for mouse clicks should be completely the same, but this should fix touch event coordinates.