Unbox QJSValue in config property map
Currently if one saves a QJSValue into the property map we save the
correct real value, but the propertymap still holds the data as a
QJSValue when retrieved.

This causes an error if a property map is moved between engines as a
QJSValue can only be in one. This can be seen in the system tray where
you can't configure entries on the first load.

By using the updateValue we can intercept updates QML and the property
map making sure only real datatypes are stored in our map.

REVIEW: 125927