Load QtQuick rendering settings from a config file and set default
Qt are slowly exposing more config options on QtQuick rendering. We want
to use this in plasmashell/screenlocker/kwin/krunner and maybe more.

The goal of this is threefold:

    Optimise when deploying on a fixed platform (raspberry pi, a
specific device, like a mobile phone or something)

    To test QSG_RENDER_LOOP=threaded on Mesa with more users, so we can
ask Qt to revisit the default. (the reason it was originally disabled is
now fixed), and it makes a *huge* performance difference

    steps towards allowing for 2D renderer fallback