Start drag with press and hold on touch events
Summary:
This patch consists of two changes:
- starting drags on press and hold
- not starting drags when the event is from a touch event

This means a user can scroll on a widget explorer with a touch screen
correctly without it immediately starting a drag, but can still drag an
item using a press and hold pattern.

BUG: 368698

Test Plan:
Scrolled in Widget Explorer
Added widget with press and hold

Surprisingly that's the only Plasma usage of DragArea.

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: mart, hein, broulik, plasma-devel, #frameworks

Tags: #plasma, #frameworks

Differential Revision: https://phabricator.kde.org/D4972