disconnect render() on window change
Summary:
when the window changes, the node will be deleted, and render()
will access an invalid pointer.
disconnect the render slot when the window change, as it
can't do anything useful until the next updatepaintnode.
Make the managedtexturenode actually do the management, as
textures were never deleted

BUG:388508
BUG:374280
BUG:365052
BUG:343576

Test Plan:
The bug is easily reproducible without the patch, wasn't
able to reproduce it anymore with it

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, broulik, ngraham, plasma-devel, #frameworks

Tags: #plasma, #frameworks

Differential Revision: https://phabricator.kde.org/D10250