Support nvidia reset notifications in QtQuickViews
Summary:
In line with changes in Qt, we support
GL_PURGED_CONTEXT_RESET_NV. On a graphical error
we recreate entire QtQuick scene contexts.

It can break custom render loops, so the change
is opt-in. This flag enables it for all plasma windows.

Test Plan:
Suspended PC, resume.
Icons on desktop are not garbled pernamently.
Unpatched kwin with Aurorae does not busy-loop forever.

Reviewers: #plasma, hein

Reviewed By: #plasma, hein

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D11960