Set kcm context property on the correct context
Summary:
One code path of kdeclarative uses a shared engine.

If we have a shared engine we have one engine with a root context and
several sub contexts. In this case it's important to set properties on
the context of our QQmlObject, not the shared root context of the
engine.

Test Plan:
I didn't see it affecting anything as one tends to only open one KCM at a time

Loaded the fonts KCM in system settings
Loaded the volume KCM via the plasma-pa applet

Reviewers: #plasma, hein

Reviewed By: #plasma, hein

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13223