[GridViewKCM] improve contrast and legibility for delegates' inline hover buttons
Summary:
Right now the delegates' inline hover buttons can be very difficult to see, because they're
(mostly) white and don't get much background contrast. A gradient appears on the bottom of
the thumbnail on hover, which provides a bit of visual differentiation, but no real contrast.

This patch improves the situation by using real buttons (with a built-in background) instead
of toolbuttons.

BUG: 395510
FIXED-IN: 5.56

Test Plan:
{F6589836, autoplay=true}

{F6589841, autoplay=true}

Reviewers: #vdg, #plasma, broulik, GB_2, ndavis

Reviewed By: #vdg, GB_2, ndavis

Subscribers: Codezela, bruns, rizzitello, ndavis, andreask, filipf, abetts, GB_2, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D18649