Enable glGetGraphicsResetStatus support by default on Qt >= 5.13
Summary:
I added support in Qt for handling Nvidia's glGetGraphicsResetStatus
flag, which reports when it has lost it's context and the visual state
needs refreshing.

As it required some changes in various other parts of Qt and Plasma this
was opt-in behind a QSurfaceFormat flag which I exposed to Plasma users
via a config option. Since Qt5.12.2 everything seems to work nicely.

This patch enables support on Qt >= 5.13 to play safe.

BUG: 364766

Test Plan: Feedback from wide test audience on the bug report.

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20087