introduce FallbackTapHandler
Summary:
This element is intended to be used as "click on empty areas
to dismiss" use case.
a MouseArea or a TapHandler are not enough for this, because Flickables
eat mouse events even tough no dragging would happen
this filters explicitly events of flickables to emit clicks when no drag
happened on them

Test Plan:
have a listview as child of the FallbackTapHandler and delegates with
a mousearea. The handler will emit tapped only when the user taps on an empty area
of the listview, but not on tap of its delegates

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22860