[ColorButton] Forward accepted signal from ColorDialog
Allows to react to the user explicitly accepting the dialog and picking a color.
While there is a currentColor vs color property on the ColorDialog, you can still end up with unexpected change signal emissions when the underlying model changes.
This signal is only ever emitted when the ColorDialog is accepted.

Differential Revision: https://phabricator.kde.org/D22964