[GridDelegate] Don't select unselected item when clicking on any of its action buttons
Summary:
In 6258f33fce3af55f7cdd3b5487cd04d8e53c0b38, I changed the delegate action button
behavior to select an unselected item when any of its buttons were clicked. In
retrospect, this was a mistake and causes various issues with items becoming selected
when they shouldn't be.

This patch reverts that part of the change to the prior behavior.

BUG: 404536
FIXED-IN: 5.62

Test Plan:
Preview an unselected LNF theme or splash screen and navigate away from the KCM; no
"unsaved changes" dialog

Click on the delete or show in folder button on an unselected wallpaper item; still works

Reviewers: broulik, #plasma

Reviewed By: broulik, #plasma

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22975