[KDeclarative] Port away from deprecated QWheelEvent methods
Summary:
This finishes the work to make kdeclarative compile with Qt 5.14,
when disabling deprecated methods.

Note that the orientation property will have the hardcoded value
Vertical (it's really surprising that EventGenerator::sendWheelEvent,
i.e. the conversion in the other direction, was hardcoding it to Horizontal)

Test Plan: Builds

Reviewers: davidedmundson, mart

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23937