Add ManagedConfigModule
Summary:
This is a new type of ConfigModule which will manage the state of
"apply" and "defaults" by itself depending on the state of the KConfigXT
objects used within the module.

This is sligthly uncomplete on the "defaults" side due to missing
facilities in ConfigModule, but at least it shows the approach works.

Reviewers: #plasma, #frameworks, mart, bport, davidedmundson

Reviewed By: #plasma, bport, davidedmundson

Subscribers: davidedmundson, bport, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24821