Introduce SettingState* elements to ease KCM writing
Summary:
This is the QML based counterpart of D27540. Unlike with the KCModule
case, this is not automatically propagated to the ConfigModules, they
will all have to be adapted to make use of it.

I provide another patch which ports a few ConfigModule to see
how it looks: D27841.

Reviewers: crossi, hchain, meven, bport, davidedmundson, mart, ngraham, #frameworks, #plasma

Reviewed By: bport

Subscribers: broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27840