Expose checkForConflictsAgainst property
The new property does the same as the property of the same name of
KKeySequenceWidget from KXmlGui. This can be helpful for users that want to
keep track of conflicts themselves such as configuration modules.  The helper
class already had such a member so we only need to expose it in the public api.