Add source to KDeclarativeMouseEvent
For hover events we don't have an explicit source in the event, but Qt
doesn't synthesise them so it's probably safe to assume any event is
real.

KDeclarativeMouseEvent is only exposed via QML so we don't need ABI
compatiability when the constructor is changed.

See https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/170