[simplekcm] Remove custom padding handling
We override the default padding of Kirigami.ScrollablePage here.

On the desktop we use whatever is defined as the default of QQC2.Control, which happens to be smaller than the default of Kirigami.ScrollablePage
On mobile we use 0, presumably to save space, but it looks bad.

Given most of the time we use SimpleKCM we use a FormLayout as central item we rarely got in the situation where the padding matters.