[abstractkcm] Set explicit padding
The current code relies on the default padding of Control, which is style dependant.

With qqc2-desktop-style and Breeeze it evaluates to 6.
With Plasma style it evaluates to 0, making it look bad.

Usually we'd use Kirigami.Units.largeSpacing here, but that's inconsistent with the current widgets style.
This is noticable in combination with the systemsettings button footer.

Since there's no trivial and style-independant way to obtain the metric from Breeze copy the value here.