[KQuickAddons] QtQuickSettings::checkBackend() for fallback to software backend
Fixes very annoying crashes of QtQuick KDE applications:
ksmserver-logout-greeter, krunner (if not running before), and others
occuring when the OpenGL driver stops being available (e.g. on upgrade).

This issue often occurs after upgrading graphic drivers, preventing reboot,
which in turn is needed for the drivers to work and not cause this crash.

No changes are required in the above applications, as they all already use
KQuickAddons::QtQuickSettings::init().

BUG: 422995
BUG: 421493
BUG: 398199
BUG: 375913
BUG: 366156
BUG: 346519