RFC: Allow enabling Qml debugging at runtime
Normally if you want to enable Qml debugging you need to build with
-DQT_QML_DEBUG or put a QQmlDebugEnabler in your code. This is cumbersome to do
for every application that one wants to debug. This adds a QQmlDebugEnabler in
a central place that can be enabled in Debug mode by setting the environment
variable "KDE_ENABLE_QML_DEBUG". This is a union of the combination of the
approach taken by plasmashell which hides it only behind an environemt variable
and another one seen in other repos to have the compile time definition for
debug builds. This is not intended as a final solution but discussion start which
one of both approaches we should go with.