Introduce Lanczos filter element
Kwin currently contains a nice way of downsampling texures in a way that
provides a more appealing result.

This patch exposes the same feature to QtQuick. Intended initial usage
is the window thumbnails, but we want it available for more users.

Performance wise, even when using PlasmaCore.WindowThumbnail as a source
and updating constantly usage of the the filter still used <1% of my
CPU. There's almost no work at all on the GUI thread.

This is based on a shader from libRetro that implements Lanczos
resampling in a single pass. This allows us to rely on pure QML
ShaderEffect for the rendering.