use Kirigami Cards for grid delegates, with visible buttons
KCM grid delegates right now have buttons that appear on hover. Hover
buttons have questionable usability, since they are not super
discoverable ant not usable at all on touch, where we currently just
make the buttons unconditionally visible, which looks a bit ugly and
heaby with the current design. For these reasons, we have been moving
away from this for some time, making list view buttons non-hover in
conjunction with a minor redesign few Frameworks releases ago. So it's
time to do the same thing for grid delegates.

Multiple rounds of VDG attempts to make the buttons always visible with
the current design or a tweaked version of it failed because this simply
clashed with the existing style. So VDG produced a new mockup of a more
card-style delegate, which has been implemented here using a Kirigami
Card for consistency.

Closes https://invent.kde.org/plasma/plasma-workspace/-/issues/2