[GridDelegate] Show inline controls for active item too
Right now the inline controls only appear for hovered items. This causes
two problems:

1. Not touch friendly since there is no concept of hover
2. It's not obvious that there are inline actions at all unless you
   think to hover over them (many users do not)

We can fix both of these problems with a fairly simple change: always
show the inline actions for the currently active item. This fixes
problem #1 by making the actions appear when a delegate is tapped, and
fixes problem #2 by providing users a hint that there are hidden hover
actions, and making the actions for the current item easier to see.