[GridDelegate] Remove lightening overlay on hover
This overlay causes the colors of the hovered item to be distorted,
which may be a problem when the content's color matters--for example
when selecting a new color scheme!

There is an outline effect on hover that differs from the selection
effect, so that should be enough to denote that an item is hovered.

Also the lightening effect is is currently hardcoded to not exist on
mobile, so let's just remove it on the desktop view to be consistent.

BUG: 438462
FIXED-IN: 5.84