[ConfigModule] Allow creating KCMs without KAboutData
This adds a new constructor that receives a KPluginMetaData instance.
This metadata contains all the information needed to replace the usage
of KAboutData.

Unfortunatly we cannot use the existing ctor that takes KPluginMetaData
since the parameters are in a different order than KPluginFactory
expects.

Internally it is checked whether the KCM has been instantiated with a
KPluginMetaData and falls back to KAboutData otherwise

If the KCM was instantiated from a KPluginMetaData and aboutData() is
called a KAboutData object is created on demand