SimpleKCM: remove custom header and footer handling
This code is completely unnecessary because the component inherits from
Kirigami.ScrollablePage, which already knows how to handle headers and
footers. Currently this code has the effect of inappropriately applying
page margins to header and footer elements. Removing the custom code
fixes that bug and results in no visible regressions.

Now KCM.SimpleKCM's only purpose is to automatically set the title and
draw lines above or below the view when it's scrollable, on the
assumption that the view as well as its headers and footers are all
using the Window background color. We should consider removing it in
KF6 and upstreaming the scroll separator functionality into
Kirigami.ScrollablePage, if needed.