KCM GridDelegate: Display a blue line when the delegate has active focus
GridDelegate currently displays nothing to indicate that it has active
focus. This means that if you move focus to a grid in a grid kcm,
there's nothing to indicate your focus is there. This changes that so
that GridDelegate draws an blue line underneath the title/subtitle in a
similar style as is used on some other controls like radio buttons.