Allow KCM kpackages to use metadata of C++ plugin
This way kpackages do not need to install the metadata files if they
are already shipped as embedded json metadata. This way we also don't
need to parse the metadata twice & keep it in memory.

By preferring the about data for the component name we ensure that we do not
change the component name with any KCMs that want to utilize the new constructor.
This is for example the case in the icons kcms.

Task: https://phabricator.kde.org/T14983