Make QML code depend less on the `kcm` context property
It allows reusing components from org.kde.kcm module in other code which
is not running in a KCM module. A marginal use-case, but a possible
one. For example, according to lxr[1] search, the Timer applet is the
only applet that uses those controls, but still it shouldn't spam logs
with meaningless 'undefined reference' messages, since mostly
(without going into applet's general usability concerns) it seems to be
working perfectly fine.

[1]: https://lxr.kde.org/search?%21v=kf5-qt5&_filestring=&_string=org.kde.kcm