[KeySequenceItem] Make keySequence «non-null», «non-undefined» and «non-""»
This property should always be of type QKeySequence and nothing else.

Unfortunately though,

1) QKeySequence is not exported in QML, and even if it were, being a
QGadget makes it virtually impossible to use as a property type because
of mind-blowing variety of internal errors from QML engine, such as:

    Invalid property assignment: unsupported type "QKeySequence*"
    TypeError: Property 'KeySequence' of object [null] is not a function
    TypeError: Type error
    Error: Cannot assign QKeySequence to QKeySequence*

2) Qt factory methods of QKeySequence are not marked to be Q_INVOKABLE
from QML side, I had to introduce one in helper class, so that we can
always construct a valid object (gadget) with an empty sequence instead
of messing with null / undefined / empty string values as before.