AbstractKCM: make footer top padding optional
Right now, AbstractKCM has explicit logic to add extra top padding to an
empty footer. This is done so that unframed QML KCMs without footers
look correct when opened in System Settings, where the System Settings
shell app itself adds buttons on the bottom of the window that require
the extra padding to be added in KCMs to look good.

However ths extra padding is not required when a QML KCM using this
component that has an empty footer doesn't have any bottom buttons added
by System Settings itself. An exmple is the Bluetooth KCM, and the GHNS
dialog also uses it.

It's not possible to automatically detect when this automatically-added
padding won't be needed so we can omit it. As a result, the only way to
let implementations opt out of it is by making it configurable using a
new public property.

We should get rid of this in KF6+Plasma 6 and do it properly by having
System Settings add the padding to its own footer, but this isn't
possible now because it would break too many other things due to
differences in the Plasma and Frameworks release schedules.