Sync header and footer separator code to SimpleKCM
AbstractKCM has complex logic to add separator lines at the top and
bottom of the page, based on the presence or absence of headers and
footers and the potential need for extra spacing on the bottom of KCMs
that have no footer and System Settings will show at least one of its
own footer buttons.

However SimpleKCM lacks much of this logic, despite having the same needs.
Instead, it has simpler logic to naïvely draw separator lines with no
provision for extra spacing. As a result, KCMs that use SimpleKCM and
become scrollable have insufficient bottom padding between the System
Settings-provided buttons and the KCM-provided separator line.

This commits syncs the code and API from AbstractKCM to SimpleKCM so
that the issues solved there are solved here too.

Code is copied rather than refactoring SimpleKCM to be based on
AbstractKCM because the trigger conditions are subtly different, we
can't add the typical padding to avoid regressions KInfoCenter's
CommandOutputKCM (see diff; we can fix this only in KF6). Such a
refactor is a separate challenge that can be undertaken later in KF6
times.

BUG: 461435
FIXED-IN: 5.101