AbstractKCM: remove KF6 TODO
This can't actually be done in KF6, as fundamentally we have to add the
extra spacing in KCM code rather than in System Settings. System
Settings can't know when the extra padding is required and when it's
not, because this requires knowledge of the KCM's own layout. Only KCMs
have that knowledge, so they need to be able to communicate this via the
API we already have.