Track hover of children
Event handling of items that filter child events changed behavior a bit
in Qt6. When going over a child it used to *not* receive an
hoverLeveEvent

while now we get an hoverleave, then the filtered child will get and
hoverenter, and finally the parent will get an hoverenter again, having
from the user code pov, an hoverleave too much, changing behavior of the
user code.
In particular, this broke the mouseover for Plasma's FolderView.

This patch tracks also the hover state for filtered children, so
containsmouse will always be true if either the item or a child has the
mouse and doesn't emit changed signals when they are not needed