keysequencehelper: clarify key naming for single-key sequences
Even though everything referred to here is a key on the keyboard, a lot
of people aren't used to seeing key names with otherwise human-readable
text. For keys like this that don't consist of a modifier key and
another key, let's clarify the situation by changing the string a bit.