kquickcontrols: Replace superfluous Text item and a semi-broken expression
The Text component wasn't here for text metrics or anything like that;
it was just a fancy way of converting a keySequence to string. And your
first thought might be "well, String() exists?" But here's a catch:
if you assign a random string to the helper.currentKeySequence property,
it might contain a key sequence whose length is 1 and its [0] element is
Qt::Key_unknown, which translates to an empty string. Luckily there is
already a purpose-made method available and exposed to QML.