kquickcontrols: Introduce private KeySequenceValidator type
This extracts the validation logic out of KeySequenceHelper and allows
us more control over the exact execution order of things. Most
importantly, this does not rely on internal message boxes and the nasty
nested event loops that come with them. Instead, KeySequenceValidator is
a little more stateful but allows the using code to indicate what should
happen if something needs handling by the user.

CCBUG: 483199