kquickcontrols: Use KeySequenceValidator in KeySequenceItem
Replace the internal logic of KeySequenceHelper with
KeySequenceValidator. This gives us control of what should happen when a
conflict has been detected, in this case we add some QtQuick dialogs to
ask those questions in a way that is compatible with QtQuick rather than
using nested event loops that cause crashes.

BUG: 483199