kquickcontrols: Fix mistakes in KeySequenceItem.pattern migration
In commit 858b7e76, the property alias `multiKeyShortcutsAllowed`
in KeySequenceItem.qml was changed from
`helper.multiKeyShortcutsAllowed` to `helper.modifierlessAllowed`.
This appears to have been an oversight in large-scale migration of
API documentation comments to QDoc.

This commit changes it back to `helper.multiKeyShortcutsAllowed`.

In commit 0bdf292e, keysequencetest.qml replaced a line
`modifierOnlyAllowed: true` with the line
`patterns: ShortcutPattern.Modifier | ShortcutPattern.Key`.

These lines mean different things: the default value
`ShortcutPattern.ModifierAndKey` (`0x4`) is not included here;
the combination of `Modifier | Key` (`0x1 | 0x2`) does not produce
the same result.

This commit changes it to the initially intended combination
`patterns: ShortcutPattern.Modifier | ShortcutPattern.ModifierAndKey`.
It conveniently overlooks the fact that this does not seem to be
used by any actual test, which is presumably why tests didn't break
in the first place.