DeclarativeDragArea: Filter mouse events according to acceptedButtons()
When starting or within a drag, childMouseEventFilter() needs to
take account of the buttons that the area expects to handle.

The symptom of this happens with plasma-addons/applets/quicklaunch,
where a right button click over a launcher icon starts a drag as
well as opening the context menu, despite acceptedButtons() being
set to Qt::LeftButton only.

BUG: 384009
FIXED-IN: 6.22